+++ frontend/src/metabase-lib/lib/queries/drills/column-filter-drill.js
@@ -1,0 +1,24 @@
+import { isa, TYPE } from "metabase/lib/types";
+
+const INVALID_TYPES = [TYPE.Structured];
+
+export function columnFilterDrill({ question, clicked }) {
+  const query = question.query();
+
+  if (
+    !question.isStructured() ||
+    !query.isEditable() ||
+    !clicked ||
+    !clicked.column ||
+    INVALID_TYPES.some(type => isa(clicked.column.base_type, type)) ||
+    clicked.column.field_ref == null ||
+    clicked.value !== undefined
+  ) {
+    return null;
+  }
+
+  const { dimension } = clicked;
+  const initialFilter = dimension.defaultFilterForDimension();
+
+  return { query, initialFilter };
+}
