+++ frontend/src/metabase/modes/components/drill/ColumnFilterDrill.jsx
@@ -1,28 +1,17 @@
 /* eslint-disable react/prop-types */
 import React from "react";
 import { t } from "ttag";
-import { TYPE, isa } from "metabase/lib/types";
 
 import FilterPopover from "metabase/query_builder/components/filters/FilterPopover";
-
-const INVALID_TYPES = [TYPE.Structured];
+import { columnFilterDrill } from "metabase-lib/lib/queries/drills/column-filter-drill";
 
 export default function ColumnFilterDrill({ question, clicked }) {
-  const query = question.query();
-  if (
-    !question.isStructured() ||
-    !query.isEditable() ||
-    !clicked ||
-    !clicked.column ||
-    INVALID_TYPES.some(type => isa(clicked.column.base_type, type)) ||
-    clicked.column.field_ref == null ||
-    clicked.value !== undefined
-  ) {
+  const drill = columnFilterDrill({ question, clicked });
+  if (!drill) {
     return [];
   }
 
-  const { dimension } = clicked;
-  const initialFilter = dimension.defaultFilterForDimension();
+  const { query, initialFilter } = drill;
 
   return [
     {
